feat: app dir sucks
