+++ next.config.mjs
@@ -8,6 +8,15 @@ const nextConfig = {
     locales: ["en-US", "zh-CN"],
     defaultLocale: "en-US",
   },
+  redirects: async () => {
+    return [
+      {
+        source: "/nominate",
+        destination: "/",
+        permanent: true,
+      },
+    ];
+  }
 };
 
 export default nextConfig;
