+++ package.json
@@ -31,10 +31,10 @@
     "hono": "4.2.7",
     "iron-session": "8.0.1",
     "luxon": "^3.4.4",
-    "next": "14.2.1",
+    "next": "^14.2.3",
     "pino-pretty": "^11.0.0",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "siwe": "2.3.2",
     "viem": "^2.9.26",
     "wagmi": "^2.5.21"
