+++ src/app/(home)/@nominateBuilder/[...catchAll]/page.tsx
@@ -1,3 +1,3 @@
 export default function Nothing () {
-  return null;
+    return null;
 }
