+++ src/app/(home)/@nominateBuilder/nominate/[userId]/loading.tsx
@@ -1,0 +1,2 @@
+import { NominateBuilderLoading } from "@/app/_components/@nominateBuilder";
+export default NominateBuilderLoading;
