+++ src/app/(home)/@nominateBuilder/nominate/[userId]/not-found.tsx
@@ -1,0 +1,2 @@
+import { NominateBuilderNotFound } from "@/app/_components/@nominateBuilder";
+export default NominateBuilderNotFound;
