+++ src/app/(home)/@nominateBuilder/nominate/[userId]/page.tsx
@@ -1,0 +1,5 @@
+import { NominateBuilder} from "@/app/_components/@nominateBuilder";
+import Link from "next/link";
+export default () => <h1>Kiwi
+    <Link href="/">Home</Link>
+</h1>
