+++ src/app/_components/@nominateBuilder/index.ts
@@ -1,0 +1,3 @@
+export { default as NominateBuilder } from './page';
+export { default as NominateBuilderLoading } from './loading';
+export { default as NominateBuilderNotFound } from './not-found';
