+++ src/app/_components/@nominateBuilder/loading.tsx
@@ -1,0 +1,8 @@
+import { DateTime } from "luxon";
+import { NominateBuilderComponent } from "./component";
+
+
+export default function NominateBuilderLoading() {
+    const date = DateTime.now().toFormat("LLL dd");
+    return (<NominateBuilderComponent loading date={date} />);
+}
