+++ src/app/_components/@searchBuilder/index.ts
@@ -1,0 +1,1 @@
+export { default as SearchBuilder } from "./page";
