--- src/app/_components/@searchNominations/index.ts
@@ -1,1 +1,0 @@
-export { default as SearchNominations } from "./page";
