+++ src/app/airdrop/@bossPointsCard/default.tsx
