+++ src/app/airdrop/@dailyStreakCard/default.tsx
