+++ src/app/airdrop/@howToPlay/default.tsx
