+++ src/app/airdrop/@nominateBuilder/default.tsx
@@ -1,3 +1,3 @@
 export default function Nothing () {
-  return null;
+    return null;
 }
