+++ src/app/airdrop/@nominateBuilder/nominate/[userId]/page.tsx
@@ -1,0 +1,2 @@
+import { NominateBuilder} from "@/app/_components/@nominateBuilder";
+export default NominateBuilder;
