+++ src/app/layout.tsx
@@ -29,6 +29,7 @@ export const metadata: Metadata = {
 
 export default function RootLayout({
   children,
+  modal,
 }: Readonly<{
   children: React.ReactNode;
 }>) {
